requirements.txt
# Gemini 2.5 Pro 视频分析工具依赖

# Google AI SDK
google-generativeai>=0.8.0

# 其他依赖
requests>=2.31.0
typing-extensions>=4.5.0